pyproject.toml
[project]
name = "ibl_alignment_gui"
version = "0.1.0"
description = "Electrophysiology and histology alignment tool"
authors = [
    {name = "International Brain Laboratory", email = "software-devs@internationalbrainlab.org"},
]
dependencies = [
  "qtpy>=2.4.1",
  "numpy",
  "pyqtgraph>=0.13.7",
  "ONE-api>=3.0.0",
  "iblutil>=1.20.0",
  "ibl-neuropixel>=1.9.0",
  "ibllib",
  "iblatlas",
  "oursin",
  "matplotlib",
  "pandas",
  "simpleitk"
]

requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project.scripts]
alignment-gui = "ibl_alignment_gui.app.main:launch_app"

[tool.pdm]
distribution = true


[dependency-groups]
dev = [
    "ruff>=0.12.10",
]
lint = ["ruff>=0.9.2"]

[tool.ruff]
line-length = 99
include = ["pyproject.toml", "ibl_alignment_gui/**/*.py", "tests/**/*.py"]
exclude = ["ibl_alignment_gui/utils/qt/adapted_axis.py", "ibl_alignment_gui/plugins/*.py"]

[tool.ruff.format]
quote-style = "single"

[tool.ruff.lint]
ignore = [
  "D100",    # Missing docstring in public module
  "D104",    # Missing docstring in public package
  "PLR2004", # Magic value used in comparison
  "PLR0913", # Too many arguments
]

select = [
  "A",    # flake8-builtins
  "B",    # flake8-bugbear
  "ICN",  # flake8-import-conventions
  "LOG",  # flake8-logging
  "PIE",  # flake8-pie
  "SIM",  # flake8-simplify
  "TID",  # flake8-tidy-imports
  "TC",   # flake8-type-checking
  "E",    # pycodestyle Error
  "W",    # pycodestyle Warning
  "F",    # Pyflakes
  "I",    # isort
  "N",    # pep8-naming
  "PL",   # pylint
  "UP",   # pyupgrade
  "D",    # pydocstyle
  "FURB", # refurb
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["D"] # pydocstyle

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.lint.isort]
known-first-party = ["brainbox*", "ibl*", "one*" ]